+++ CHANGELOG.md
@@ -21,6 +21,7 @@ This will break the projects developed on precedent versions.
 * ADD: getStemmerDesc, getPosConverterDesc to Morpho
 * ADD: singular 2 plural to Arabic, English and French (a testing function)
 * ADD: singular to dual to Arabic (a testing function)
+* ADD: splitToSentences, tokenize and filterStopWords to Morpho  API
 
 ## Version 0.7.0 (2018-01-01)
 * NEW: French support (Info, Lang, Trans, Morpho)